[Queue] added queue to kail project

+++ Inc/queue_fifo.h
@@ -12,17 +12,14 @@
 	
 	typedef struct Queue {
 	uint8_t Array_size;
-	int *First_array_element;
-	int *Current_element;
-	int *Next_element;
-	int *Last_element;
+	uint8_t *First_array_element;
+	uint8_t *Current_element;
+	uint8_t *Next_element;
+	uint8_t *Last_element;
 	uint8_t Space_left;
-	
 } Queue_data;
-	
+
 Queue_data* Initialize_queue(uint8_t Queue_size);
 void Destroy_Queue(Queue_data *Queue_name);
-int Add_element(Queue_data *Queue_name, uint8_t Element);
+uint8_t Add_element(Queue_data *Queue_name, uint8_t Element);
 uint8_t Dequeue_element(Queue_data *Queue_name);
-
-
